perf: move some items to core module and others to users module
