+++ src/components.d.ts
@@ -13,8 +13,8 @@ declare module 'vue' {
     CarbonSun: typeof import('~icons/carbon/sun')['default']
     CarbonWarning: typeof import('~icons/carbon/warning')['default']
     Footer: typeof import('./components/TheFooter.vue')['default']
-    README: typeof import('./components/README.md')['default']
-    TheFooter: typeof import('./components/TheFooter.vue')['default']
+    README: typeof import('./core/components/README.md')['default']
+    TheFooter: typeof import('./core/components/TheFooter.vue')['default']
   }
 }
