+++ src/core/components/README.md
