+++ src/core/layouts/DefaultLayout.vue
