+++ src/core/modules/i18n.ts
@@ -6,9 +6,9 @@ import { createI18n } from 'vue-i18n'
 //
 // Don't need this? Try vitesse-lite: https://github.com/antfu/vitesse-lite
 const messages = Object.fromEntries(
-  Object.entries(import.meta.globEager('../../locales/*.y(a)?ml')).map(([key, value]) => {
+  Object.entries(import.meta.globEager('../../../locales/*.y(a)?ml')).map(([key, value]) => {
     const yaml = key.endsWith('.yaml')
-    return [key.slice(14, yaml ? -5 : -4), value.default]
+    return [key.slice(17, yaml ? -5 : -4), value.default]
   })
 )
