+++ src/core/modules/nprogress.ts
