+++ src/core/modules/pinia.ts
