+++ src/core/pages/README.md
