+++ src/core/pages/[...all].vue
