+++ src/core/pages/index.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
-import { useUserStore } from '~/stores/user'
+import { useUserStore } from '~/users/stores/user'
 
 const user = useUserStore()
 const name = ref(user.savedName)
 
 const router = useRouter()
 const go = () => {
-  if (name.value) router.push(`/hi/${encodeURIComponent(name.value)}`)
+  if (name.value) router.push(`/users/greet/${encodeURIComponent(name.value)}`)
 }
 
 const { t } = useI18n()
